--- a/pyproject.toml
+++ b/pyproject.toml
@@ -21,12 +21,13 @@ dependencies = [
     "fhir.resources==7.1.0",
     "numpy==1.26.4",
     "orjson==3.9.13",
-    "pandas>=2.2.0",
+    "pandas[performance]>=2.2.0",
     "pyarrow==15.0.0",
     "pydantic==2.6.1",
     "pydantic_core==2.16.2",
     "tzdata",
     "python-dateutil",
+    "joblib",
 ]
 
 [project.optional-dependencies]