Different approach to increasing ingestion speed.

This PR uses `joblib` to parallelise the ingestion across resources.
The current two time sinks are:
 * reading in the data & mapping files (not much we can do about that)
 * the 'create_dictionary' function, likely because of all the iterating over row/columns using `apply` going on.

Without parallelisation, current time is:
```
Patient took 0.88 seconds to convert 67 rows. 
Encounter took 0.74 seconds to convert 67 rows. 
2 resources not created due to validation errors. Errors saved to encounter_errors.csv
Observation took 1.81 seconds to convert 2194 rows. 
Condition took 1.07 seconds to convert 461 rows. 
Total time: 4.4922334590228274
```

With parallelisation one, the speed up is:
```
Patient took 0.78 seconds to convert 67 rows. 
Encounter took 0.73 seconds to convert 67 rows. 
2 resources not created due to validation errors. Errors saved to encounter_errors.csv
Observation took 1.71 seconds to convert 2194 rows. 
Condition took 0.86 seconds to convert 461 rows. 
Total time: 3.185586916981265
```
(readout edited for easy comparison)