@pipliggins 2024-08-07
While I think there might be more speed increases by re-writing some of the dictionary creation, I'm going to see how slow it gets when creating the 'full' dengue mapping files, in comparison to the time taken to read in the data files, before worrying about it too much more.